refactor: rename and consolidate modules related to legacy analytics reports (#4912)
